Выносит функции-хелперы в util.js
