--- js/createElementFromTemplate.js
@@ -1,7 +1,0 @@
-export default (templateString) => {
-  const template = document.createElement(`div`);
-
-  template.innerHTML = templateString.trim();
-
-  return template;
-};
