+++ js/main.js
@@ -1,4 +1,4 @@
-import renderScreen from './renderScreen.js';
+import {renderScreen} from './util.js';
 import introScreen from './intro.js';
 
 renderScreen(introScreen);
