--- js/renderScreen.js
@@ -1,6 +1,0 @@
-const mainTemplate = document.querySelector(`main.central`);
-
-export default (element) => {
-  mainTemplate.innerHTML = ``;
-  mainTemplate.appendChild(element);
-};
